Migrated random_indices_within_bin to utils.array_index_manipulations
