+++ halotools/empirical_models/phase_space_models/subhalo_based_models/tests/test_subhalo_selection_kernel.py
@@ -55,71 +55,3 @@ def test_full_index_selection():
     assert set((fake_satellite_objids[2], )) <= set((5, 6))
     assert set((fake_satellite_objids[3], )) <= set((9, 10))
     assert set(fake_satellite_objids[4:]) <= set((12, 15, 16))
-
-
-def test_random_indices_within_bin1():
-    binned_multiplicity = np.array([1, 1, 1])
-    desired_binned_occupations = np.array([2, 3, 4])
-
-    result = ssk.random_indices_within_bin(
-        binned_multiplicity, desired_binned_occupations, seed=43)
-    correct_result = np.repeat(np.arange(len(binned_multiplicity)), desired_binned_occupations)
-    assert np.all(result == correct_result)
-
-
-def test_random_indices_within_bin2():
-    binned_multiplicity = np.array([0, 1, 1])
-    desired_binned_occupations = np.array([2, 3, 4])
-
-    with pytest.raises(ValueError) as err:
-        result = ssk.random_indices_within_bin(
-            binned_multiplicity, desired_binned_occupations, seed=43)
-    substr = "Input ``binned_multiplicity`` array must contain "
-    assert substr in err.value.args[0]
-    substr = "min_required_subs_per_bin = 1"
-    assert substr in err.value.args[0]
-
-    with pytest.raises(ValueError) as err:
-        result = ssk.random_indices_within_bin(
-            binned_multiplicity, desired_binned_occupations,
-            seed=43, min_required_subs_per_bin=13)
-    substr = "min_required_subs_per_bin = 13"
-    assert substr in err.value.args[0]
-
-
-def test_random_indices_within_bin3():
-    """ The purpose of this test is to scour the set of possible inputs for
-    edge cases that may not be covered by the hand-tailored tests.
-    """
-
-    num_bins = 10
-    max_binned_multiplicity = 5
-    seed_list = np.arange(100).astype(int)
-
-    test_exists_with_more_subs_than_sats = False
-    test_exists_with_more_sats_than_subs = False
-    test_exists_with_more_sats_than_subs_alt_corner_case = False
-    for seed in seed_list:
-        with NumpyRNGContext(seed):
-            binned_multiplicity = np.random.randint(1, max_binned_multiplicity, num_bins)
-            desired_binned_occupations = np.random.randint(0, 2*max_binned_multiplicity, num_bins)
-
-        result = ssk.random_indices_within_bin(
-            binned_multiplicity, desired_binned_occupations, seed=43)
-
-        if binned_multiplicity.sum() > desired_binned_occupations.sum():
-            test_exists_with_more_subs_than_sats = True
-            if np.any(desired_binned_occupations > binned_multiplicity):
-                assert len(result) > len(set(result))
-                test_exists_with_more_sats_than_subs_alt_corner_case = True
-        if desired_binned_occupations.sum() > binned_multiplicity.sum():
-            test_exists_with_more_sats_than_subs = True
-
-        assert len(result) == desired_binned_occupations.sum()
-        assert np.all(result >= 0)
-        assert np.all(result <= binned_multiplicity.sum())
-
-    # Verify that our chosen seed_list covered both of the following cases
-    assert test_exists_with_more_subs_than_sats is True
-    assert test_exists_with_more_sats_than_subs is True
-    assert test_exists_with_more_sats_than_subs_alt_corner_case is True
